Refactor model and tests change openPack function
Remove the initialize method that added 15 cards to the base Pack model.
This will change based on the type of pack. Create a mock TestPack that
extends the base pack object and fills it with cards on init for use in
testing. Do this going forward for all pack implementations.

Players begin with currentPack set to null, openPack function checks for
null or empty pack or it will not open. Added and updated tests to reflect
this behavior.

Refactored tests to check for cids in comparisons and not use instanceof