Add activePackEmpty and allActivePacksEmpty functions
Added methods to player and draft, which will return true if the player's active pack is empty or if all the active packs of all players are empty respectively. Implemented left to right pack passing method based on what the packNumber in the draft is set to.

TODO:

Make the packNumber property go up automatically when the new packs are opened.